feat(Req 01): :sparkles: Implements the Queue class within queue.py
The class Queue initializes with the creation of an empty list within the __init__ method, done with self.queue = list(). After this step, another four methods are defined. __len__ returns the length of the list created in __init__. enqueue pushes a new value to the queue list with the append(value) call. dequeue removes the first element from the list and returns the element that just got removed, and for last, the method search which verifies first if the index is a valid integer, if not, it raises an IndexError, else, it will return the element present on the index argument.