feat(Req 02): :sparkles: Implements the function txt_importer within file_management.py
The function firstly checks if the path_file call in the parameter is an existent file. If it doesn't exist, it writes in the stderr "Arquivo {path_file} não encontrado\n". After validating whether the files exists, the function tests for the file extension. If the file doesn't have a ".txt" extension, it returns "Formato inválido\n" on the stderr. After all the tests, if they pass, the function returns a list with the lines from the .txt file. Also, for an expected output, the lines were formatted with the .strip() method for taking off the \n character at the end of each line.