feat(Req 04): :sparkles: Implements remove function within file_process.py
The remove function is responsible of checking if the actual queue has any elements. If the queue is empty, it returns "Não há elementos\n" on stdout. If the queue has any element, it procedes to call the dequeue() method and returns "Arquivo {path_file} removido com sucesso\n" on the stdout.