feat(Req 05): :sparkles: Implements file_metadata function within file_process.py
The file_metadata checks if any data is existent within the queue when accessing a specific index passed by the parameter position. If it returns an IndexError (it means, a index that ain't present in the queue), it prints in the stderr "Posição inválida\n". Else, it returns the data at the informed position within the queue list.