feat(Req 06): :sparkles: Implements exists_word function within word_search.py
The existis_word function iterates over the queue list testing if the word passed as an argument exists within the lines existing inside the list "linhas_do_arquivo". If true, the list "ocorrencias" will be populated with dicts containing the index of the line where the word was found. If the "ocorrencias" iteration fails to encounter any match, the function just returns an empty list.